made a new class with SW & Our new Interface
